ccgx: Add a child device for ASYMMETRIC devices to represent the other firmware
Some CCGX devices are symmetric, which means we can write from either A/B
partition of the firmware and switch to that in one operation. On some devices
that are described as asymmetric we have to update the 'other' partition first
then switch to the degraded bootloade mode, *then* use that to flash the 'normal'
partition, then finally reboot.

This child device allows us to install FW2 then FW1 in the same transaction
without asking the user to manually re-run the transaction with a post-update
console message.

The reason this works is that two FuInstallFasks are added to the queue as both
components in the cab file match the parent **and** child devices.

One the parent has been installed, the device replugs and the parent and child
swap around, and the other install task is ran -- on the same device but in a
different mode. Once the second install has completed the device again rebots
into runtime mode.

To the client, this asymmetric device looks like:

    │
    └─ThinkPad USB-C Dock Gen2 PD Main Controller:
      │   Device ID:           4ebf0e4fea1a48cf47c2774a812c681fd0256cc4
      │   Summary:             PD-IC Primary Firmware
      │   Current version:     0.0.33
      │   Vendor:              Cypress Semiconductor (USB:0x04B4)
      │   Install Duration:    1 minute
      │   GUIDs:               2c4d586a-a2a6-5957-9427-768add598121 ← USB\VID_04B4&PID_521A&REV_0000
      │                        ebd96940-fa49-5cb1-b1a1-aa2d4dea1416 ← USB\VID_04B4&PID_521A
      │                        5108e726-44a6-5eca-bd1a-fbbdabbe7bc1 ← USB\VID_04B4&PID_521A&SID_1F00&APP_6D64
      │                        31d41b9b-361a-50ad-8bea-8dad3339fa04 ← USB\VID_04B4&PID_521A&SID_1F00&APP_6D64&MODE_FW2
      │   Device Flags:        • Updatable
      │                        • Requires AC power
      │                        • Device will not re-appear after update completes
      │                        • Cryptographic hash verification is available
      │                        • Device stages updates
      │                        • Device can recover flash failures
      │
      └─ThinkPad USB-C Dock Gen2 PD Backup Controller:
            Device ID:         61e5898da990afcbb8c7e866f42fd6470f52591c
            Summary:           PD-IC Backup Firmware
            Current version:   0.0.33
            Vendor:            Cypress Semiconductor (USB:0x04B4)
            Install Duration:  1 minute
            GUIDs:             5108e726-44a6-5eca-bd1a-fbbdabbe7bc1 ← USB\VID_04B4&PID_521A&SID_1F00&APP_6D64
                               f6b5141d-eeb0-5ff1-b8de-ee4f02b222b7 ← USB\VID_04B4&PID_521A&SID_1F00&APP_6D64&MODE_FW1
            Device Flags:      • Updatable
                               • Requires AC power
                               • Install to parent device first
                               • Cryptographic hash verification is available
                               • Device stages updates
                               • Device can recover flash failures

Only mark the child device as updatable whenever in the primary firmware, or
when in the backup FW and have seen the primary at least once.

This prevents us doing `RECOVERY -> PRIMARY -> RECOVERY` on a device with
initially corrupt firmware. Techically this sequence is correct -- but probably
not what the user actually wants.